Edb quality functionality

+++ main.js
@@ -1,47 +1,50 @@
-// var btnDelete = document.querySelector('#btn-delete');
-// var btnDownvote = document.querySelector('#btn-downvote');
 // var btnGenius = document.querySelector('#btn-genius');
 // var btnPlausible = document.querySelector('#btn-plausible');
 // var btnQuality = document.querySelector('#btn-quality');
 var btnSave = document.querySelector('#btn-save');
 // var btnStar = document.querySelector('#btn-star');
-// var btnStarred = document.querySelector('#btn-star');
+// var btnStarred = document.querySelector('#btn-starred');
 // var btnSwill = document.querySelector('#btn-swill');
-// var btnUpvote = document.querySelector('#btn-upvote');
 var inputBody = document.querySelector('#input-body');
 // var inputQuality = document.querySelector('#input-quality');
 // var inputSearch = document.querySelector('#input-search')
 var inputTitle = document.querySelector('#input-title');
 var cardArea = document.querySelector('.section-bottom');
 var ideasArray = []
-// for (i = 0; i > ideasArray.length; i++) {
+var id = Date.now();
+var qualitiesArray = ["Swill", "Plausible", "Genius"];
 
+
+// for (i = 0; i > ideasArray.length; i++) {
 //   (ideaCard + i) = document.querySelector(`idea-card-${i}`);
 //   (ideaCard + i).addEventListener('blur', updateCard);
 // }
-var id = Date.now();
+
 
 // btnGenius.addEventListener('click', );
 // btnPlausible.addEventListener('click', );
 // btnQuality.addEventListener('click', toggleArrow);
-btnSave.addEventListener('click', addIdea);
-cardArea.addEventListener('click', toggleStar);
 // btnSwill.addEventListener('click', );
-
+btnSave.addEventListener('click', addIdea);
+cardArea.addEventListener('click', handleCardButtons);
+cardArea.addEventListener('click', toggleStar);
 cardArea.addEventListener('focusout', handleFocusOut);
 cardArea.addEventListener('keydown', handleTextEdit);
-cardArea.addEventListener('click', handleCardButtons);
-inputTitle.addEventListener('keyup', handleSaveBtn);
-
 inputBody.addEventListener('keyup', handleSaveBtn);
+inputTitle.addEventListener('keyup', handleSaveBtn);
 window.addEventListener('DOMContentLoaded', repopulateIdeasArray);
 
-// handleBottom() {
-// 	if (e.)
-// 		make if functions to target every button/image
-// }
+
 function handleCardButtons(e) {
-  deleteCard(e);
+  if (e.target.id === 'btn-upvote') {
+    upvoteQuality(e);
+  }
+  if (e.target.id === 'btn-downvote') {
+    downvoteQuality(e);
+  }
+  if (e.target.id === 'btn-delete') {
+    deleteCard(e);
+  }
 }
 
 function addIdea(e) {
@@ -70,7 +73,7 @@ function addCard(object) {
         </div>
         <footer>
           <img class="img-upvote" src="images/upvote.svg" alt="upvote" id="btn-upvote">
-          <p class="quality-text">Quality: ${object.quality}</p>
+          <p class="quality-text">Quality: ${qualitiesArray[object.quality]}</p>
           <img class="img-downvote" src="images/downvote.svg" alt="downvote" id="btn-downvote">    
         </footer>
       </article>`);
@@ -86,6 +89,35 @@ function repopulateIdeasArray() {
   for (i = 0; i < ideasArray.length; i++) {
     addCard(ideasArray[i]);
   }
+}
+
+function upvoteQuality(e) {
+  e.target.closest('.idea-card');
+  var index = findIdeaIndex(e);
+  if (ideasArray[index].quality < qualitiesArray.length - 1) {
+  var newQuality = ideasArray[index].quality + 1;
+  ideasArray[index].updateQuality(newQuality);
+  qualityDisplay(e);
+  ideasArray[index].saveToStorage(ideasArray);
+  }
+}
+
+function downvoteQuality(e) {
+  e.target.closest('.idea-card');
+  var index = findIdeaIndex(e);
+  if (ideasArray[index].quality > 0) {
+  var newQuality = ideasArray[index].quality - 1;
+  ideasArray[index].updateQuality(newQuality);
+  qualityDisplay(e)
+  ideasArray[index].saveToStorage(ideasArray);
+  }
+}
+
+function qualityDisplay(e) {
+  var qualityDisplay = e.target.closest('.idea-card').querySelector('.quality-text');
+  var index = findIdeaIndex(e);
+  var ideaQuality = qualitiesArray[ideasArray[index].quality];
+  qualityDisplay.innerText = `Quality: ${ideaQuality}`
 }
 
 function deleteCard(e) {
@@ -113,8 +145,6 @@ function handleFocusOut(e) {
     ideasArray[index].updateIdea(newTitle, newBody);
     ideasArray[index].saveToStorage(ideasArray);
     console.log("index:", index);
-    // console.log(title.innerText);
-    // console.log(body.innerText);
   }
 }
 
@@ -139,15 +169,12 @@ function handleTextEdit(e) {
  } 
 
 function handleSaveBtn() {
-  console.log("inputTitle", inputTitle.value);
-  console.log("inputBody", inputBody.value);
 	btnSave.disabled = !inputTitle.value || !inputBody.value;
 }
+
   // console.log(title);
   //return enter key saves changes
   //assign the new fields to the property values on DOM
   //get the objects with the changes
   //update the array to include new changes
   //pass the array to local storage udpateIdea() to update data model
-
-
